Problem-1
You are given a sorted (either in the increasing or in the decreasing order) sequence of numbers, ending with a -1. You can assume that are at least two numbers before the ending -1. You have to output the number of distinct elements in the sorted sequence. Don’t use arrays in the code.